+++ testLongDouble.c
@@ -1,0 +1,20 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+int main(void)
+{
+
+//    long int data;
+//    scanf("%li", &data);
+//    printf("%li", data);
+//    
+//    
+//    printf("Answer is : %li \n", 2147483647);
+//    printf("Answer is : %li", 2147483649);
+
+//	long double data;
+//	scanf("%Lf", &data);
+//	printf("%Lf",data);
+    
+   return 0;
+}
